+++ audit_hcahps_funcs.py
@@ -1,4 +1,4 @@
-﻿# CAHPS Auditor - HCAHPS-specific functions
+# CAHPS Auditor - HCAHPS-specific functions
 # Copyright (C) 2026 HST Pathways. All rights reserved.
 # Originally developed by Tyler Brock. This copyright notice, including
 # authorship credit to Tyler Brock, must be preserved in all copies,