+++ audit_lib_funcs.py
@@ -1,4 +1,4 @@
-﻿# CAHPS Auditor
+# CAHPS Auditor
 # Copyright (C) 2026 HST Pathways. All rights reserved. Developed by Tyler Brock. 
 # This copyright notice must be preserved in all copies, modifications, and derivative works of this software.
 import re