+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿# Changelog (Releases)
+# Changelog (Releases)
 
 All notable changes to this project will be documented in this file.